requirements.txt
# Telegram Video Downloader Bot Dependencies

# Telegram Bot API wrapper
python-telegram-bot>=20.0

# Video downloader
yt-dlp>=2023.0.0

# Environment variable management
python-dotenv>=1.0.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
hypothesis>=6.0.0
